Move getStsCredential to iam.go for consistency
