+++ pkg/controller/account/account_controller.go
@@ -12,7 +12,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/iam"
 	"github.com/aws/aws-sdk-go/service/organizations"
-	"github.com/aws/aws-sdk-go/service/sts"
 	"github.com/aws/aws-sdk-go/service/support"
 	"github.com/go-logr/logr"
 	awsv1alpha1 "github.com/openshift/aws-account-operator/pkg/apis/aws/v1alpha1"
@@ -766,50 +765,6 @@ func CreateUserAccessKey(reqLogger logr.Logger, client awsclient.Client, userNam
 	}
 
 	return result, nil
-}
-
-// getStsCredentials returns sts credentials for the specified account ARN
-func getStsCredentials(reqLogger logr.Logger, client awsclient.Client, iamRoleName string, awsAccountID string) (*sts.AssumeRoleOutput, error) {
-	// Use the role session name to uniquely identify a session when the same role
-	// is assumed by different principals or for different reasons.
-	var roleSessionName = "awsAccountOperator"
-	// Default duration in seconds of the session token 3600. We need to have the roles policy
-	// changed if we want it to be longer than 3600 seconds
-	var roleSessionDuration int64 = 3600
-	// The role ARN made up of the account number and the role which is the default role name
-	// created in child accounts
-	var roleArn = fmt.Sprintf("arn:aws:iam::%s:role/%s", awsAccountID, iamRoleName)
-	reqLogger.Info(fmt.Sprintf("Creating STS credentials for AWS ARN: %s", roleArn))
-	// Build input for AssumeRole
-	assumeRoleInput := sts.AssumeRoleInput{
-		DurationSeconds: &roleSessionDuration,
-		RoleArn:         &roleArn,
-		RoleSessionName: &roleSessionName,
-	}
-
-	assumeRoleOutput := &sts.AssumeRoleOutput{}
-	var err error
-	for i := 0; i < 100; i++ {
-		time.Sleep(500 * time.Millisecond)
-		assumeRoleOutput, err = client.AssumeRole(&assumeRoleInput)
-		if err == nil {
-			break
-		}
-	}
-	if err != nil {
-		// Log AWS error
-		if aerr, ok := err.(awserr.Error); ok {
-			reqLogger.Error(aerr,
-				fmt.Sprintf(`New AWS Error while getting STS credentials, 
-					AWS Error Code: %s, 
-					AWS Error Message: %s`,
-					aerr.Code(),
-					aerr.Message()))
-		}
-		return &sts.AssumeRoleOutput{}, err
-	}
-
-	return assumeRoleOutput, nil
 }
 
 func (input SRESecretInput) newSTSSecret() *corev1.Secret {
